requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
Django==3.1.6
django-dotenv==1.4.2
djangorestframework==3.12.2
idna==3.3
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
urllib3==1.26.9
